Split time-tests common library
Applied clang-format-9 and added a README.md
